7/27/18 1

+++ BlazorLib2/Classes/SubClasses/rect.cs
@@ -7,12 +7,12 @@ namespace BlazorLib2.Classes.SubClasses
 {
     public class rect:strokeBase
     {
-        public int x { get; set; }
-        public int y { get; set; }
-        public int rx { get; set; }
-        public int ry { get; set; }
-        public int width { get; set; }
-        public int height { get; set; }
+        public double x { get; set; }
+        public double y { get; set; }
+        public double rx { get; set; }
+        public double ry { get; set; }
+        public double width { get; set; }
+        public double height { get; set; }
         public string style { get; set; }
         public string fill { get; set; }
